+++ Editor/FramePropertyDrawer.cs
@@ -8,6 +8,7 @@ namespace AnimExpressEditor
 	public class FramePropertyDrawer : PropertyDrawer
 	{
 		private const float DURATION_FIELD_WIDTH = 40f;
+		private const float TEXTURE_FIELD_WIDTH = 60f;
 
 		public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
 		{
@@ -20,17 +21,24 @@ namespace AnimExpressEditor
 			var duration = property.FindPropertyRelative("duration");
 
 			Rect spriteRect = position;
-			spriteRect.width = position.width - AnimationExpressEditorUtils.COPY_BUTTON_WIDTH - DURATION_FIELD_WIDTH;
+			spriteRect.width = position.width - TEXTURE_FIELD_WIDTH - AnimationExpressEditorUtils.COPY_BUTTON_WIDTH - DURATION_FIELD_WIDTH;
+
+			Rect textureRect = position;
+			textureRect.width = TEXTURE_FIELD_WIDTH;
+			textureRect.x += spriteRect.width;
 
 			Rect copyButtonRect = position;
 			copyButtonRect.width = AnimationExpressEditorUtils.COPY_BUTTON_WIDTH;
-			copyButtonRect.x += spriteRect.width;
+			copyButtonRect.x += spriteRect.width + TEXTURE_FIELD_WIDTH;
 
 			Rect durationRect = position;
 			durationRect.width = DURATION_FIELD_WIDTH;
-			durationRect.x += spriteRect.width + AnimationExpressEditorUtils.COPY_BUTTON_WIDTH;
+			durationRect.x += spriteRect.width + TEXTURE_FIELD_WIDTH + AnimationExpressEditorUtils.COPY_BUTTON_WIDTH;
 
 			EditorGUI.PropertyField(spriteRect, sprite, GUIContent.none);
+
+			GUI.DrawTexture(textureRect, AssetPreview.GetAssetPreview(sprite.objectReferenceValue), ScaleMode.ScaleToFit);
+
 			if (GUI.Button(copyButtonRect, EditorGUIUtility.IconContent("Clipboard")))
 			{
 				GUIUtility.systemCopyBuffer = $"\"{sprite.objectReferenceValue.name}\"";